1: Fix author taglet registration
The registration of the custom 'author' failed because of an InvocationException thrown in Taglet.register(...). Added catch clause to identify exceptions happening during registration. The problem was that Java also registers a taglet with the name 'author' for a taglet object that can not be downcastet to net.digitalid.utility.taglets.Taglet. Unfortunately, Java taglets and custom-built taglets do not have any common interfaces (Java Taglets use com.sun.tools.doclets.internal.toolkit.taglets.Taglet, custom-built taglets use com.sun.tools.doclets.Taglet) so we can only cast to 'Object'. Shouldn't be an issue though, because we remove it anyway if it is not null in the next step.