Fixed issues in subclass and builder generator.
Added new annotations @GenerateNoBuilder and @GenerateNoSubclass to mark classes so that the specific generators ignore them.
Fixed constructor generation. Instead of representing fields we use the constructor parameters + generated fields in there now.
Fixed generation of interface implementations.

Currently still missing is the generation of the pre- and post-condition of TYPE_USE annotations.