Extended Subclass and Builder for conversion
Added a collectValues() to the subclass and a static restore() method to the builder, which can be used by the conversion mechanism.
In collectValues(), a set*() method is called for each representative field. Likewise, in the builder, a get*() method is called for every
representative field. Two things are not cleanly solved yet:
First of all, the restore method does not seem to work for our Lists (needs more testing). Second, non-representative fields that are required for the builder, will not be set. This might lead to unexpected run-time exceptions when required fields are not set. We should at least output a warning if a type cannot be restored solely through its representative fields.

Also: Inverted the default for the generation of builders and subclasses. Now, no Subclasses or Builders are generated, unless the type is annotated with @GenerateSubclass and @GenerateBuilder.